Start merging bot script
